fix(animation): Remove animation also from scene
Instead of only removing the animation from the "animation-manager", we now also remove it from the
related scene with a new action "removeAnimationFromEverywhere".

fix #85
