+++ src/actions/index.js
@@ -3,6 +3,7 @@ import * as selectors from '../selectors/index.js'
 import * as utils from '../utils/index.js'
 
 export * from './timeline.js'
+export * from './animation.js'
 
 /*
  *
@@ -216,64 +217,6 @@ export const resetAllFixtures = () => {
       // Reset the propeties of the fixture in the state & batch
       dispatch(resetFixtureProperties(fixture.id))
     })
-  }
-}
-
-/*
- * Add a animation
- */
-export const addAnimation = animation => ({
-  animation,
-  type: constants.ADD_ANIMATION
-})
-
-/*
- * Update a animation
- */
-export const setAnimation = animation => ({
-  animation,
-  type: constants.SET_ANIMATION
-})
-
-/*
- * Start the playback of a animation
- */
-export const runAnimation = animationId => ({
-  animationId,
-  type: constants.RUN_ANIMATION
-})
-
-/*
- * Remove an animation
- */
-export const removeAnimation = animationId => ({
-  animationId,
-  type: constants.REMOVE_ANIMATION
-})
-
-
-/*
- * Add a keyframe to an animation
- */
-export const addKeyframe = (animationId, keyframeStep, keyframeProperty, keyframeValue) => ({
-  animationId,
-  keyframeStep,
-  keyframeProperty,
-  keyframeValue,
-  type: constants.ADD_KEYFRAME
-})
-
-/*
- * Add keyframes to an animation
- */
-export const addKeyframes = (animationId, keyframeStep, keyframeProperties) => {
-  return (dispatch, getState) => {
-    Object.entries(keyframeProperties).map(keyframeProperty => {
-      const [property, value] = keyframeProperty
-
-      dispatch(addKeyframe(animationId, keyframeStep, property, value))
-    })
-
   }
 }
